requirements.txt
python-binance>=1.0.19
websocket-client>=1.7.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
python-telegram-bot>=20.8
ccxt>=4.2.0
python-dotenv>=1.0.1
pytest>=8.0.0
matplotlib>=3.8.0
scipy>=1.12.0
requests>=2.31.0
asyncio>=3.4.3